+++ dwds/CHANGELOG.md
@@ -12,6 +12,7 @@
 - Include an optional param to `Dwds.start` to indicate whether it a Flutter app
   or not.
 - Pre-warm expression compiler cache to speed up Flutter Inspector loading.
+- Remove `ChromeProxyService.setExceptionPauseMode()`.
 
 ## 16.0.1